Revert back to import statements within individual functions.
Satrajit Ghosh writes re: nipype:
'for the function nodes, you cannot have module level inputs. they have to be imported inside the function.
the reason for this is that functions may be executed on any computer and we simply send the function as a string.
thus whatever code is injected into the workspace needs to be self contained.'


Former-commit-id: 402f861492651efa79c5fbe182b972c0bfeb01cc