Matlab code to compute likelihood values.
1) - 3) are for training the values, with 1) being the main function. 4) is for computing likelihood for a given subject after the distributions have been trained

Learning distributions, from training data:
1)

[aDepth, aCurv] = parameterizeDepthAndCurvature(norm_depths,curvature,label_borders_sulci,Sulci)

This function computes the EM estimates for depth and curvature distributions. Both outputs are size Kx6, where K is the number of classes. Columns 1 to 3 are respective mean, standard deviation and weight for class 'Non-label boundary'. Columns 4 to 6 are the same thing for class 'label boundary'. K=3 is currently hard-coded in subfunction.

2)

a = learnDistributionParameters(feature,Sulci,LabelBoundarySulci,x)

This is a subfunction of 1). Note that the lines inside the if-statement are just to add noise to curvature values, and should be removed once the curvature maps are fixed.

3)

[clMeans clSigmas W] = fitKNormalsOnHistogramEM(data,x)

This is the EM-method called by 2), currently with hard-coded K=3. The outputs are means, standard deviations (sigmas) and weights (W), each of size K x 1.

//

Finally for computing the likelihood for any subject:

4)

L = compParameterizedLikelihoodGaussianMixture(aDepth,aCurv,Sulci,norm_depths,curvature)

given the outputs of 1), computes likelihoods based on input values and 'Sulci'.


Former-commit-id: 92134911a21567fa7c3e137db842c84f43f4ca37