Add iterative thresholding to find_endpoints().
The new function decreases the likelihood threshold iteratively,
favoring faster segmentation along higher likelihood directions,
and the first segment to reach the boundary establishes an endpoint.

Including an X region of previously segmented vertices,
expanding this into R, P, and N to reduce propagation along the walls
as opposed to along the length of branches, did not work.

The function is able to extract very reasonable endpoints for the Y-shaped
test case (fold 11 in Twins-2-1 left hemisphere), but this does not
generalize to the other folds well.