Include file to optionally filter vertices through which to connect fundi.
The area file is used in the pipeline to reduce the likelihoods of vertices that have large area,
and are therefore less likely to be on high-density, highly curved portions of fold where fundi
are more likely to be.

The parameters for find_outer_anchors() and connect_points_erosion() are exposed in
extract_fundi() and therefore can be set in the nipype pipeline.


Former-commit-id: 04c537467ccc5a47a8a1e988896fc40c02d5c649