Major overhaul of the nipype pipeline.
All workflows called by a single workflow (mbFlow) rather than sequentially run.
This was not possible before because both hemispheres would need to finish labeling
before filling cortical gray matter volume with labels. This 2nd step is now run twice.

The new registration workflow computes affine (and nonlinear) transform(s) to standard space,
and the transforms are applied to the surface mesh coordinates for the tables.
The nonlinear transforms will be used to label subcortical volumes...

Functions and files have been renamed and moved around in wake of this reorganization.

To do: the area values are per-vertex in the tables rather than per-structure.
To do: subcortical labeling workflow.


Former-commit-id: bc20168b7ba479289538c3b4f2bd69df8c858c9d