Revise apply_affine_transform() to include swap_dim1dim2() function.
This function can be used to swap first two dimensions before and after applying transform.
In an exchange between Satrajit Ghosh and Brian Avants, they determined that swapping the 1st and 2nd
set of coordinates is required before and after applying a transform to point data
(but not to nifti volume data).


Former-commit-id: a68eeda2ee3a4a5547cf8f579edd6679556cd40f